grub: Extend module scope to actual problem (#1229341)
Original texts and meta-data focus on 'mentioning' a GRUB/splash.xpm.gz
workaround which is not the most important thing we have to say.  The
point is that GRUB to GRUB2 migration has to be done manually and it
applies to all GRUB installations.

This commit fixes the relevancy logic and updates all rexts accordingly.

Text correction by jhornice.

Signed-off-by: Petr Stodulka <pstodulk@redhat.com>