bind/configuration: parse config file correctly (#1449231)
Previously the regular expressions have been used for parsing
of the configuration file, that could lead into
 - broken configration file after fixing of the content
 - immense time consuming searching of specific parts of the file,
   when regexp doesn't cover some valid strings of the config file,
   even when BIND it accepts.
As a solution it is now used mini parser that correctly finds
searched parts of the configuration file by fixing functions. In the
same way it is used for safe removing of comments from configuration
files (which was probably correct even before). Use of regexp has been
kept for search of specific keys, as these are safe after remove
of all comments and it is improbable that it could appears
in different meaning then now.