use correct combination of risks and exit statuses
With the new PA the exit codes will be changed and some combinations
of module results and risks will not be possible. Mainly, the risks
has to be set iff the exit code of the module check is FAIL (102),
otherwise the result of the module will be presented as error.

For that reson:
 - dovecot
     - set $RESULT_FAIL instead of $RESULT_FIXED
 - ReplacedPackages, openssh/sysconfig
     - use log_info instead of log_slight_risk