openssh/sysconfig: correct logic and regexps
Previously the module assumed that only problem inside the config file
would be using of the *export* command on lines as:
    export VARIABLE=VALUE
and in case that files like that don't exist, exit with PASS status.
But the file can contains ANY executable code, so the check was wrong.

Now the PASS or FIXED result is achieved only in case when we can
ensure the content of the file produce same configuration on both
systems (original and upgraded).

Additionaly, in case we cannot ensure the file is correct, it is copied
into dirtyconf and admin is instructed to verify (&modify) it move it
to cleanconf if the new configuration should be applied automatically
on the terget system.

- additionaly it is corrected syntax error in previous solution
- texts have been updated appropriately