Support for assessment format version
We want to assume that other tools will be able to "understand"
output data. For this purpose modules will provide a new key inside
the properties.ini file: "ast_fmt_version". This will contain version
in X.Y format, with semantic:
 - X will be bumped, when "something" has been changed or removed,
     so other tools need to be checked, whether they are still
     compatible with modules or not
 - Y will be bumped, when just "something" new has been added, but
     tools should be still compatible with current version

The "something" will be specified later in docs. Currently we can
say, that it is everything, that is parsed, read, required by other
tools for their correct functionality, but it doesn't include e.g.
  specific pre-upgrade scripts
  specific post-upgrade scripts
  parts of outputs in reports
  ...
But we can say, that it includes e.g.:
  dirtyconf directory
  hooks directory
  kickstart/RHRHEL7rpmlist_replaced file; that is used when kickstart
                                          is generated
  ...

Resolves #31