initscripts/control: update preset file
Update bundled preset file to newer one we expect will be used in
RHEL 7.5. In addition add there hypervfcopy service as enabled even
when it is not part of the preset file on target system. The service
is enabled automatically when corresponding device is detected. This
change suppress unwanted message in report.

Resolves #53.