system/requirements: check dmidecode and dmesg before use
Some architectures don't contain dmidecode utility at all and
it does not have to be installed. So to suppress the error msg
just check the utility exists. In addition I check even existence
of the *dmesg* utility, however it should be installed always on
the system as part of the util-linux-ng (on RHEL 6) which is required
by the PA/PAM. From this point, there is possibly the dead code
as the info from the /proc/meminfo file will be never used, but
keeping that just to have seatbelts in case we would not get info
from the dmesg.

Additionaly one warning message has been removed, as it was
irrelevant for the user (in 99.9% cases will not give any useful info
for the user) and the error message has been lowered to warning as
it was not an error (here it could gives useful info in some cases,
but usually you will not see the message at all).

Fixes: #40