tomcat: copy original webapps directory to the new path
Original APP_WEB_HOME* paths contains symlinks to the paths where
the apps should be stored. But the tomcat6 RPM is now removed before
the post-upgrade is processed, so the migration couldn't be done
correctly. Use now the defualt paths where webapps are usually
stored as a temporary solution.

FIXME: Correct paths should be set based on the original settings
of the system.

Fixes: #55