Drop the InterVariants module
The module was broken, producing bunch of FP messages. As the set
of dark-matrix data contains just repositories in the standard
compose, all other rpms (extras, various layered products, ...)
are marked as rpms that should not be installed under the installed
RHEL 6 variant (Server, CompudeNode,...). Including PA rpms.

It's not easy to fix it and the module is purely informative. So we
are dropping it instead.